requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.2.3
blinker==1.4
blis==0.7.4
brotlipy==0.7.0
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==2.9.2
cymem==2.0.5
decorator==4.4.2
defusedxml==0.6.0
en-core-web-sm==3.0.0
entrypoints==0.3
Flask==1.1.2
gast==0.4.0
gingerit==0.8.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
langdetect==1.0.8
language-check==1.1
language-tool-python==2.5.2
Markdown==3.3.3
MarkupSafe==1.1.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==4.7.6
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
nltk==3.5
notebook==6.2.0
numpy==1.19.2
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pathy==0.4.0
pickleshare==0.7.5
pip==20.3.3
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
pyenchant==3.2.0
Pygments==2.7.4
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.17.3
PySocks==1.7.1
pyspellchecker==0.6.0
python-dateutil==2.8.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scipy==1.5.2
Send2Trash==1.5.0
setuptools==52.0.0.post20210125
six==1.15.0
smart-open==3.0.0
spacy==3.0.3
spacy-legacy==3.0.1
srsly==2.4.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
thinc==8.0.1
tornado==6.1
tqdm==4.57.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
yarl==1.5.1
zipp==3.4.0
